Checks scopes to guard against leaks.
One could leak a dependency by providing a scoped dependency from a
component with a shorter lifetime to one with a longer lifetime. We
detect this by ensuring the injected dependence's scope is always at
least as long as the dependency is it injected into. Unscoped
dependencies are always allowed as they don't have any particular
lifetime.

Fixes #188