Add a warning for applying a scope on a generated component function or property as it will have no effect.
It should be placed on a @Provides function or @Inject constructor
instead.

Note: due to the limitation with the KAPT back-end it will miss cases
where the property is annotated itself instead of it's getter. ex:
`@MyScope abstract val myProp` is missed but
`@get:MyScope abstract val myProp` is caught. This can be improved once
the KAPT back-end is removed.

Partially addresses #238