Don't reuse symbols between rounds
Reusing symbols between rounds is not supported and throws an error on
ksp2. Unfortunately this means we need to use `getSymbolsWithAnnotation`
instead of our custom functions as they are currently the only way to
access updated deferred symbols. We still keep a list of the symbol
names so that we can run a final processing pass on the last round.

See https://github.com/google/ksp/issues/1854