Fix `FileAlreadyExistsException`
Classes keep being processed multiple times and throw a `FileAlreadyExistsException` in multi round scenarios. The function `getSymbolsWithAnnotation()` returns deferred symbols twice for some reason in the next round.

The fix is only applied for classes and not functions yet, because there the issue didn't occur. More care should also be taken for functions, because the qualified name alone is not an indicator to distinguish functions as functions be overloaded and have different parameters.