[WIP] rewrite code generation
The original code gen would often explode in code size. The reason for
this is it would generate the entire chain of provided dependencies in
each instance and rely on an optimization pass to combine common code.
When this optimization missed cases it would cause code bloat.

The new method is to optimize in the opposite direction. i.e generate a
common provider for each dependency and then inline cases where the
provider is only used once. This guarantees only 1 instance of each
dependency creation instead of n, even if the optimization pass misses
anything.